I edited delete ticket functionality
Only the creator of the ticket can delete it.Otherwise an error message will appear

+++ help-desk-api/src/main/java/com/api/help_desk_api/services/implementation/TicketServiceImpl.java
@@ -101,10 +101,14 @@ public class TicketServiceImpl implements TicketService {
     }
 
     @Override
-    public void deleteTicket(int ticketId) {
+    public void deleteTicket(int ticketId,int userId) {
         Ticket ticket = ticketRepository.findById(ticketId)
         .orElseThrow(() -> 
         new TicketNotFoundException("Ticket not found"));
+        
+        if(userId != ticket.getUser().getId()) {
+            throw new TicketNotFoundException("Error!Only the ticket creator can delete it");
+        }
 
         ticketRepository.delete(ticket);
     }
